Update code to deal with new Marketplace nav Refactor - rename regions to pages Refactor - move code from marketplace app object to base page object
