--- marketplacetests/marketplace/regions/base_region.py
@@ -1,12 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-from marketplacetests.marketplace.app import Marketplace
-
-
-class BaseRegion(Marketplace):
-
-    def __init__(self, marionette):
-        Marketplace.__init__(self, marionette)
-        self.wait_for_page_loaded()
